Give machine gunners a 6.1-tile base range
Support fractional range in authoritative stats and weapon profiles and keep the client mirror aligned. Other weapon ranges remain unchanged. Add a replay-derived movement/combat regression across 32 approach variants; the old 6.0 range reproduces return fire, while 6.1 clears the entrenched rifleman unharmed.

+++ server/crates/rules/src/balance/stats.rs
@@ -12,7 +12,7 @@ pub const TANK_TRAP_CLUSTER_ATTACK_RADIUS_TILES: f32 = 4.0;
 pub struct UnitStats {
     pub hp: u32,
     pub dmg: u32,
-    pub range_tiles: u32,
+    pub range_tiles: f32,
     pub cooldown: u32,
     pub speed: f32,
     pub sight_tiles: u32,
@@ -56,7 +56,7 @@ pub struct BuildingStats {
     pub foot_h: u32,
     pub build_ticks: u32,
     pub dmg: u32,
-    pub range_tiles: u32,
+    pub range_tiles: f32,
     pub cooldown: u32,
 }
